+++ lumin/data_processing/file_proc.py
@@ -17,18 +17,19 @@ def save_to_grp(arr:np.ndarray, grp:Group, name:str) -> None:
 
 def fold2foldfile(df:pd.DataFrame, out_file:h5py.File, fold_idx:int,
                   cont_feats:List[str], cat_feats:List[str], targ_feats:Union[str,List[str]], targ_type:str,
-                  misc_feats:List[str]=[], wgt_feat:Optional[str]=None) -> None:
+                  misc_feats:Optional[List[str]]=None, wgt_feat:Optional[str]=None) -> None:
     '''Save fold data into foldfile group'''
     grp = out_file.create_group(f'fold_{fold_idx}')
     save_to_grp(np.hstack((df[cont_feats].values.astype('float32'), df[cat_feats].values.astype('float32'))), grp, 'inputs')
     save_to_grp(df[targ_feats].values.astype(targ_type), grp, 'targets')
     if wgt_feat is not None: save_to_grp(df[wgt_feat].values.astype('float32'), grp, 'weights')
-    for f in misc_feats: save_to_grp(df[f].values, grp, f)  
+    if misc_feats is not None:
+        for f in misc_feats: save_to_grp(df[f].values, grp, f)  
 
 
 def df2foldfile(df:pd.DataFrame, n_folds:int, cont_feats:List[str], cat_feats:List[str],
                 targ_feats:Union[str,List[str]], savename:Union[Path,str], targ_type:str,
-                strat_key:str=None, misc_feats:List[str]=[], wgt_feat:Optional[str]=None):
+                strat_key:str=None, misc_feats:Optional[List[str]]=None, wgt_feat:Optional[str]=None):
     '''Convert dataframe into foldfile'''
     savename = str(savename)
     os.system(f'rm {savename}.hdf5')