+++ lumin/nn/models/helpers.py
@@ -34,7 +34,7 @@ class Embedder():
         cat_szs = None
         # Get cardinalities
         for fld_id in range(len(fy)):
-            tmp_max = fy.get_df(pred_name=None, targ_name=None, wgt_name=None, n_folds=1, fold_idx=0, inc_inputs=True,
+            tmp_max = fy.get_df(pred_name=None, targ_name=None, wgt_name=None, n_folds=1, fold_idx=fld_id, inc_inputs=True,
                                 verbose=False, suppress_warn=True)[cat_names].max().values.astype(int)
             if cat_szs is None: cat_szs = tmp_max
             else:               cat_szs = np.maximum(cat_szs, tmp_max)