+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Master - targeting V0.2
 
 ## Important changes
+
 - Residual mode in `FullyConnected`:
     - Identity paths now skip two layers instead of one to align better with [arXiv:1603.05027](https://arxiv.org/abs/1603.05027)
     - In cases where an odd number of layers are specified for the body, the number of layers is increased by one
@@ -32,6 +33,9 @@
 - Added `from_results` and `from_save` classmethods for `Ensemble`
 - Added option to `SWA` to control whether it only updates on cycle end when paired with an `AbsCyclicalCallback`
 - Added helper class `Embedder` to simplify parsing of embedding settings
+- Added parameters to save and configure plots to `get_nn_feat_importance`, `get_ensemble_feat_importance`, and `rf_rank_features`
+- Added classmethod for `Model` to load from save
+- Added experimental export to Tensorflow Protocol Buffer
 
 ## Removals
 
@@ -48,10 +52,12 @@
 - Fixed feature selection in examples using a test size of 0.8 rather than 0.2
 - Fixed crash when no importnat features were found by `rf_rank_features`
 - Fixed rule-of-thumb for embedding sizes from max(50, 1+(sz//2)) to max(50, (1+sz)//2)
+- Fixed cutting when saving plots as pdf
 
 ## Changes
 
 - Moved `on_train_end` call in `fold_train_ensemble` to after loading best set of weights
+- Replaced all mutable default arguments
 
 ## Depreciations